requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
jinja2==3.1.4
python-multipart==0.0.9
itsdangerous==2.2.0
python-dotenv==1.0.1
bcrypt==4.2.0
sqlalchemy==2.0.32
pymupdf==1.24.9
redis==5.0.8
rq==2.6.1
requests==2.32.3
markdown==3.6
regex==2024.7.24